Get Task closer to working.
This commit fixes Task#attempt. It also removes Eval#attempt and
adds the Xor.catching method.

Eval#attempt is not going to work since Eval is not guaranteed
to be lazy. For example, given Funtion0[Eval[A]], it's not
possible to get a Function0[Eval[Xor[Throwable, A]]] instance
using Eval#attempt (since the construction of Eval[A] might
itself throw an Exception).

Since Eval#attempt wasn't required for the PR I removed it.
An alternate strategy would to be to create a subtype of Eval
which everything except Now extends (e.g. DeferEval), add the
method there, and use that instead.